Replace "blue-starred bellies" with "tattooed arms"
I personally find the blue-starred bellies example silly sounding -- especially if the reader isn't familiar with the Dr. Suess book it's based on.

I'd suggest either changing it to "tattooed arms" as in this PR, or some other example. (like the one Tim had originally, of gold-plated vs diamond-studded -- that example seems fine as well)