Edited for typos and clarity
Specifically, the following lines were edited (this is the before text):

"The quality of these arguments will vary widely with respect to their quality and strength."

"Unique ID - It is important to guarantee that each and every Claim can be referenced within an from outside of the system via a simple identifier not tied to the other changeable attributes."

"The most mundane problem is that performance."

"The following is a guide to the types of actions that a Curator may choose to execute in the effort to maintain the debates canonical and constructive."

"Whether Debaters are misunderstand what has previously been said, are too lazy to read the points that have already been made, or just want to say it in their own words, duplication is rampant in online debates."

"If the platform provides support for this type of structure, than most cases for a Split Claim Action will actually be a conversion to a Multi-premise Claim, which entails creating the separate, distinct Premise Claims, and joining them together into the original Multi-premise Claim."
 
"The usual text search capabilities should be enabled to provide users to locate any Claim that matches their interests on a best-effort basis."

"These are all interesting opportunities to browse and discover what interest other users of the platform."

"Often before the user is allowed to answer the question, the system returns several guesses as to what they were trying to ask, and asks the user to first verify if one of the items presented resolves their issue."

"Nevertheless, there is still the possibility that individuals with bad intentions can still perform a type of "Denial of Service" attack by creating duplicate Claims with inflammatory language repeatedly so that users will see them until Curators have time to clean them up."

I have not edited the following line because I am not sure what a good way to do so would be. By itself, there are no issues, but it doesn't make sense in the context of the example given:

"Imagine now that the claim regarding military intervention was determined to be mostly false, while the claim regarding medical aid was considered mostly true. The medical intervention Claim would then be used as an Argument against the general Claim, whereas the military intervention Claim would support it."